move util_prometheus to util
