+++ tests/integration/telemetry/util.go
@@ -19,15 +19,19 @@ package telemetry
 
 import (
 	"context"
+	"fmt"
 	"sort"
+	"time"
 
 	"github.com/prometheus/common/model"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"istio.io/istio/pkg/config/mesh"
 	"istio.io/istio/pkg/test"
+	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/cluster"
 	"istio.io/istio/pkg/test/framework/components/prometheus"
+	"istio.io/istio/pkg/test/util/retry"
 )
 
 // PromDiff compares a query with labels to a query of the same metric without labels, and notes the closest matching
@@ -89,7 +93,7 @@ func PromDiff(t test.Failer, prom prometheus.Instance, cluster cluster.Cluster, 
 	}
 }
 
-// PromDump gets all of the recorded values for a metric by name and generates a report of the values.
+// PromDump gets all the recorded values for a metric by name and generates a report of the values.
 // used for debugging of failures to provide a comprehensive view of traffic experienced.
 func PromDump(cluster cluster.Cluster, prometheus prometheus.Instance, query prometheus.Query) string {
 	if value, err := prometheus.Query(cluster, query); err == nil {
@@ -99,7 +103,7 @@ func PromDump(cluster cluster.Cluster, prometheus prometheus.Instance, query pro
 	return ""
 }
 
-// Get trust domain of the cluster.
+// GetTrustDomain return trust domain of the cluster.
 func GetTrustDomain(cluster cluster.Cluster, istioNamespace string) string {
 	meshConfigMap, err := cluster.Kube().CoreV1().ConfigMaps(istioNamespace).Get(context.Background(), "istio", metav1.GetOptions{})
 	defaultTrustDomain := mesh.DefaultMeshConfig().TrustDomain
@@ -118,4 +122,41 @@ func GetTrustDomain(cluster cluster.Cluster, istioNamespace string) string {
 	}
 
 	return cfg.TrustDomain
+}
+
+// QueryPrometheus queries prometheus and returns the result once the query stabilizes
+func QueryPrometheus(t framework.TestContext, cluster cluster.Cluster, query prometheus.Query, promInst prometheus.Instance) (string, error) {
+	t.Helper()
+	t.Logf("query prometheus with: %v", query)
+
+	val, err := promInst.Query(cluster, query)
+	if err != nil {
+		return "", err
+	}
+	got, err := prometheus.Sum(val)
+	if err != nil {
+		t.Logf("value: %s", val.String())
+		return "", fmt.Errorf("could not find metric value: %v", err)
+	}
+	t.Logf("get value %v", got)
+	return val.String(), nil
+}
+
+func ValidateMetric(t framework.TestContext, cluster cluster.Cluster, prometheus prometheus.Instance, query prometheus.Query, want float64) {
+	t.Helper()
+	err := retry.UntilSuccess(func() error {
+		got, err := prometheus.QuerySum(cluster, query)
+		t.Logf("%s: %f", query.Metric, got)
+		if err != nil {
+			return err
+		}
+		if got < want {
+			return fmt.Errorf("bad metric value: got %f, want at least %f", got, want)
+		}
+		return nil
+	}, retry.Delay(time.Second), retry.Timeout(time.Second*20))
+	if err != nil {
+		PromDiff(t, prometheus, cluster, query)
+		t.Fatal(err)
+	}
 }
